requirements.txt
django
djangorestframework
django-rest-auth
django-rest-auth[with_social]
djangorestframework-jwt
django-cors-headers